# Simplify & improve performance of PrettyContextBuilder