@wayilau 2018-11-04
i've registred CLA, i will take 'make test' tomorrow.