@luksa 2018-11-04
@wayilau are you sure this really improves performance? Without benchmarking this, it's impossible to know. 

IMHO, instead of using buffer, this code could be simplified and its performance would also most likely improve (I'm thinking about removing the `space` variable and adding the space into `s`. Also, removing the usage of `sprintf` would definitely boost performance.