@wayilau 2018-11-05
@luksa thanks reply me.  i couldnot say it really improves performance whthout benchmarking this, i will make this code simplified for you suggestion.