@luksa 2018-11-05
@wayilau I played around with this and ran a benchmark. Here are the results:
```
=== RUN   TestBenchmark
Original version:              7.022284207s
Concat, no space var:          4.673069881s
Concat, no space, no sprintf:  1.976600523s
Buffer, no space, no sprintf:  2.39237122s
--- PASS: TestBenchmark (16.07s)
```
Using a Buffer does appear to be slower than just using plain old string concatenation.