@wayilau 2018-11-05
@luksa i've modify all from your guide.
/retest