@luksa 2018-11-05
@wayilau please change the title of this PR to something like "Simplify & improve performance of PrettyContextBuilder". 

You also need to sign the CLA, as instructed here: https://github.com/kubernetes-incubator/service-catalog/pull/2455#issuecomment-435328034