@wayilau 2018-11-07
> Can you add the benching code so we can review it from time to time in the future?

yes, i 've added one.