@wayilau 2018-11-07
i 've rebase my commit for email chainged.  
/retest