@luksa 2018-11-07
/lgtm

I'm just not sure if @MHBauer wanted you to add the benchmark in the test itself or somewhere else.