@MHBauer 2018-11-07
go test has built in support for a benchmark, so I was hoping to see that.

https://golang.org/pkg/testing/#hdr-Benchmarks
https://dave.cheney.net/2013/06/30/how-to-write-benchmarks-in-go
https://lk4d4.darth.io/posts/bench/