@MHBauer 2018-11-07
fairly straightforward translation of the bench looks like 
```golang
var bResult string

func BenchmarkPCB(b *testing.B) {
	pcb := NewContextBuilder(ServiceInstance, "Namespace", "Name", "877")

	b.ResetTimer()
	var s string
	for i := 0; i < b.N; i++ {
		s = pcb.String()
	}

	bResult = s
}
```
and allows you to use nice tools
```console
$ benchcmp old.txt new.txt
benchmark          old ns/op     new ns/op     delta
BenchmarkPCB-4     666           166           -75.08%
```